New separate interface for ManipulationState.
