--- patterns/state/manipulator_state/states/mixins/parent_state_mixin.dart
@@ -1,1 +1,0 @@
-
